+++ examples/basic_usage.js
@@ -1,10 +1,15 @@
-var Promise = require('promise');
 var eve = require('../index');
 
-var transport = new eve.transport.DistribusTransport();
+// create two agents
+var agent1 = new eve.Agent('agent1');
+var agent2 = new eve.Agent('agent2');
+
+// create a transport and connect both agents
+var transport = new eve.transport.LocalTransport();
+agent1.connect(transport);
+agent2.connect(transport);
 
 // agent 1 listens for messages containing 'hi' or 'hello' (case insensitive)
-var agent1 = new eve.Agent('agent1');
 agent1.receive = function (from, message) {
   console.log(from + ' said: ' + message);
 
@@ -13,14 +18,9 @@ agent1.receive = function (from, message) {
 };
 
 // agent 2 listens for any message
-var agent2 = new eve.Agent('agent2');
 agent2.receive = function (from, message) {
   console.log(from + ' said: ' + message);
 };
-
-// connect both agents to the transport
-agent1.connect(transport);
-agent2.connect(transport);
 
 // send a message to agent 1
 agent2.send('agent1', 'Hello agent1!');