+++ examples/pubnub.js
@@ -8,18 +8,18 @@ var transport = new eve.transport.PubNubTransport({
 
 // agent 1 listens for messages containing 'hi' or 'hello' (case insensitive)
 var agent1 = new eve.Agent('agent1');
-agent1.on(/hi|hello/i, function (from, message) {
+agent1.receive = function (from, message) {
   console.log(from + ' said: ' + message);
 
   // reply to the greeting
   this.send(from, 'Hi ' + from + ', nice to meet you!');
-});
+};
 
 // agent 2 listens for any message
 var agent2 = new eve.Agent('agent2');
-agent2.on(/./, function (from, message) {
+agent2.receive = function (from, message) {
   console.log(from + ' said: ' + message);
-});
+};
 
 // connect both agents to the transport
 var connected1 = agent1.connect(transport);